Adds an option to calculate the estimated value of the portfolio
Our mechanism is triggered when -P parameter is passed. We try to read the portfolio configuration
file, which is a simple JSON object which has the coin symbol as keys and the amount of the possesed
coins as value. We lightly modified the fetching/parsing mechanism to render the new value.